Initialized project and files
This commit contains the project prototype. This prototype contains a script that allows the user to mix two colors together based on their hue values. However, this script stills needs to be implemented in game. I also created a basic movement and animation controls for the player that can be uses with keyboard, gamepad, or mobile control schemes.